# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "代理订阅售卖系统后端 API。统一信封 {code,message,data}；鉴权 Bearer \u003caccess_token\u003e。",
        "title": "NanoCloud API",
        "contact": {},
        "version": "1.0"
    },
    "host": "api.example.com",
    "basePath": "/api/v1",
    "paths": {
...
        "/coupons/check": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "交易"
                ],
                "summary": "优惠券试算",
                "parameters": [
                    {
                        "description": "请求",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CouponCheckReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/resp.Body"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.CouponCheckResp"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/resp.Body"
                        }
                    }
                }
            }
        },
...
        "/orders": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "交易"
                ],
                "summary": "订单列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "状态",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页条数",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/resp.Body"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/resp.Page"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "交易"
                ],
                "summary": "创建订单",
                "parameters": [
                    {
                        "type": "string",
                        "description": "幂等键",
                        "name": "Idempotency-Key",
                        "in": "header"
                    },
                    {
                        "description": "请求",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CreateOrderReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/resp.Body"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.OrderResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/orders/{order_no}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "交易"
                ],
                "summary": "订单详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "订单号",
                        "name": "order_no",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/resp.Body"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.OrderDetailResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/orders/{order_no}/cancel": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "交易"
                ],
                "summary": "取消订单",
                "parameters": [
                    {
                        "type": "string",
                        "description": "订单号",
                        "name": "order_no",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/resp.Body"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.OrderResp"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/resp.Body"
                        }
                    }
                }
            }
        },
        "/orders/{order_no}/checkout": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "交易"
                ],
                "summary": "收银台",
                "parameters": [
                    {
                        "type": "string",
                        "description": "订单号",
                        "name": "order_no",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "请求",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CheckoutReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/resp.Body"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.CheckoutResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/plans": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "交易"
                ],
                "summary": "套餐列表",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/resp.Body"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "list": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/definitions/model.PlanResp"
                                                    }
                                                }
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.CheckoutReq": {
            "type": "object",
            "required": [
                "method"
            ],
            "properties": {
                "method": {
                    "type": "string"
                }
            }
        },
        "model.CheckoutResp": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "expire_in": {
                    "type": "integer"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "model.CouponCheckReq": {
            "type": "object",
            "required": [
                "code",
                "period",
                "plan_id"
            ],
            "properties": {
                "code": {
                    "type": "string"
                },
                "period": {
                    "type": "string",
                    "enum": [
                        "month",
                        "quarter",
                        "half_year",
                        "year",
                        "onetime"
                    ]
                },
                "plan_id": {
                    "type": "integer"
                }
            }
        },
        "model.CouponCheckResp": {
            "type": "object",
            "properties": {
                "discount_amount": {
                    "type": "number"
                },
                "pay_amount": {
                    "type": "number"
                },
                "valid": {
                    "type": "boolean"
                }
            }
        },
        "model.CreateOrderReq": {
            "type": "object",
            "required": [
                "period",
                "plan_id"
            ],
            "properties": {
                "coupon_code": {
                    "type": "string"
                },
                "period": {
                    "type": "string",
                    "enum": [
                        "month",
                        "quarter",
                        "half_year",
                        "year",
                        "onetime"
                    ]
                },
                "plan_id": {
                    "type": "integer"
                }
            }
        },
...
        "model.OrderDetailResp": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "balance_used": {
                    "type": "number"
                },
                "coupon_code": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "discount_amount": {
                    "type": "number"
                },
                "order_no": {
                    "type": "string"
                },
                "paid_at": {
                    "type": "string"
                },
                "pay_amount": {
                    "type": "number"
                },
                "pay_method": {
                    "type": "string"
                },
                "period": {
                    "type": "string"
                },
                "plan_name": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "model.OrderResp": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "created_at": {
                    "type": "string"
                },
                "discount_amount": {
                    "type": "number"
                },
                "order_no": {
                    "type": "string"
                },
                "pay_amount": {
                    "type": "number"
                },
                "period": {
                    "type": "string"
                },
                "plan_name": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
        "model.PlanPrices": {
            "type": "object",
            "properties": {
                "half_year": {
                    "type": "number"
                },
                "month": {
                    "type": "number"
                },
                "onetime": {
                    "type": "number"
                },
                "quarter": {
                    "type": "number"
                },
                "year": {
                    "type": "number"
                }
            }
        },
        "model.PlanResp": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "device_limit": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "prices": {
                    "$ref": "#/definitions/model.PlanPrices"
                },
                "sort": {
                    "type": "integer"
                },
                "speed_limit": {
                    "type": "integer"
                },
                "traffic_gb": {
                    "type": "integer"
                }
            }
        },
...
        "resp.Body": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "resp.Page": {
            "type": "object",
            "properties": {
                "list": {},
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}